// template/react-starter-kit/src/.generated/api/model/adminTermDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export interface AdminTermDto {
  id: string;
  code: string;
  title: string;
  isRequired: boolean;
  sortOrder: number;
  version: string;
  content: string;
  /** @nullable */
  publishedAt: string | null;
  isPublished: boolean;
  isDraft: boolean;
  createdAt: string;
  updatedAt: string;
}